+++ packages/backend/src/routes/api/v1/history.test.ts
@@ -1,5 +1,6 @@
 import express from "express";
 import request from "supertest";
+import { isCanvasModerator } from "@/services/discordGuildService";
 import {
   deletePixelHistoryEntries,
   getPixelHistory,
@@ -10,6 +11,11 @@ import { historyRouter } from "./history";
 vi.mock("@/services/historyService", () => ({
   deletePixelHistoryEntries: vi.fn(),
   getPixelHistory: vi.fn(),
+}));
+
+vi.mock("@/services/discordGuildService", () => ({
+  isCanvasModerator: vi.fn(),
+  isCanvasAdmin: vi.fn(),
 }));
 
 const createApp = ({ authenticated = false, moderator = false } = {}) => {
@@ -81,6 +87,7 @@ describe("History route tests", () => {
       responseBody as Awaited<ReturnType<typeof getPixelHistory>>,
     );
 
+    vi.mocked(isCanvasModerator).mockResolvedValueOnce(true);
     const app = createApp({ authenticated: true, moderator: true });
     const response = await request(app)
       .post("/api/v1/canvas/9/pixel/history?x0=1&y0=2&x1=3&y1=4")
@@ -126,6 +133,7 @@ describe("History route tests", () => {
       responseBody as Awaited<ReturnType<typeof getPixelHistory>>,
     );
 
+    vi.mocked(isCanvasModerator).mockResolvedValueOnce(true);
     const app = createApp({ authenticated: true, moderator: true });
     const response = await request(app)
       .post("/api/v1/canvas/9/pixel/history?x0=1&y0=2&x1=3&y1=4")
@@ -157,6 +165,7 @@ describe("History route tests", () => {
   });
 
   it("returns 400 when both includeColors and excludeColors are provided", async () => {
+    vi.mocked(isCanvasModerator).mockResolvedValueOnce(true);
     const app = createApp({ authenticated: true, moderator: true });
 
     const response = await request(app)
@@ -176,6 +185,7 @@ describe("History route tests", () => {
   });
 
   it("deletes history entries for a moderator", async () => {
+    vi.mocked(isCanvasModerator).mockResolvedValueOnce(true);
     const app = createApp({ authenticated: true, moderator: true });
     vi.mocked(deletePixelHistoryEntries).mockResolvedValueOnce(undefined);
 
@@ -197,7 +207,7 @@ describe("History route tests", () => {
   it("returns 403 when deleting history without moderator permissions", async () => {
     const app = createApp({ authenticated: true, moderator: false });
     vi.mocked(deletePixelHistoryEntries).mockResolvedValueOnce(undefined);
-
+    vi.mocked(isCanvasModerator).mockResolvedValueOnce(false);
     const response = await request(app)
       .delete("/api/v1/canvas/1/pixel/history")
       .set("X-TestUserId", "1")
